inline the table type declaration
